% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{address1}
\alias{address1}
\title{Origin/Gender discrimination and strongly negative mediatic exposure
(information technologist)}
\format{
A data frame with 3684 rows and 11 variables
}
\usage{
data(address1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between December 2008
and January 2009. A candidate is defined by the variables ("gender","origin",
"mediaexp"). The variable "reput" creates a sample separation.

\itemize{
\item\bold{offer:} add number.
\item\bold{date:} date of the application.
\item\bold{sentorder:} order in which the application was sent.
\item\bold{callback:} TRUE if there was a non negative callback.
\item\bold{fname:} first name (forename).
\item\bold{lname:} last name (family name, surname).
\item\bold{origin:} all the candidates are French, the origin is suggested by the
name, F = France, M = Morocco.
\item\bold{gender:} Woman or Man.
\item\bold{city:} candidate location.
\item\bold{reput:} reputation of the city, P = privileged, U = Unprivileged.
\item\bold{mediaexp:} strong negative mediatic exposure, Yes or No.
}
}
\references{
Duguet E., Gray D., L'Horty Y., Du Parquet L, Petit P., 2020. Labor market
effects of urban riots:an experimental assessment. Papers in Regional
Science, 99:787-806.
}
\keyword{datasets}
