% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_workflow.R
\name{compare_Seurat_clusterings}
\alias{compare_Seurat_clusterings}
\title{Returns the adjusted Rand index for the active identities in two Seurat objects}
\usage{
compare_Seurat_clusterings(seurat_obj1, seurat_obj2)
}
\arguments{
\item{seurat_obj1}{A Seurat object}

\item{seurat_obj2}{A Seurat object}
}
\value{
The adjusted Rand index for the active identities of each of the objects.
compare_Seurat_clusterings(seurat_obj1, seurat_obj2)
}
\description{
Given two Seurat objects, returns the adjusted Rand index for the active identities of each of the objects.
}
