% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_workflow.R
\name{compute_knockoff_filter}
\alias{compute_knockoff_filter}
\title{Returns the genes selected by the knockoff filter}
\usage{
compute_knockoff_filter(seurat_obj, cluster1, cluster2, q, return_all = FALSE)
}
\arguments{
\item{cluster1}{The Idents of the cluster of interest in seurat_obj1}

\item{cluster2}{The Idents of the cluster of interest in seurat_obj2}

\item{q}{The desired rate to control the FDR at}

\item{return_all}{Determines if the returned object will contain all genes or just the selected genes.}

\item{seurat_obj1}{A Seurat object}

\item{seurat_obj2}{A Seurat object}
}
\value{
The adjusted Rand index for the active identities of each of the objects.
}
\description{
Given two Seurat objects, returns the  the genes selected by the knockoff filter and their W statistics.
}
\examples{
compute_knockoff_filter(seurat_obj1, seurat_obj2, 1, 2, 0.05)
}
