% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zipoisson.R
\name{estimate_zi_poisson}
\alias{estimate_zi_poisson}
\title{Random data generation for the zero-infalted Poisson distribution
with Poisson parameter lambda and zero proportion prop.zero.}
\usage{
estimate_zi_poisson(data)
}
\arguments{
\item{data}{The data to estimate parameters from.}
}
\value{
Maximum likelihood estimators of for the zero-inflated Poisson
distribution
}
\description{
Given data, computes the maximum likelihood estimators
for the zero-infalted Poisson distribution.
}
