% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{FindClustersCallback}
\alias{FindClustersCallback}
\title{Runs a typical Seurat workflow on a Seurat object (up to dimensionality reduction and clustering).}
\usage{
FindClustersCallback(
  seurat_obj,
  resolution_start = 0.8,
  reduction_percentage = 0.2,
  num_clusters_start = 20,
  dims = 1:10,
  algorithm = "louvain",
  assay = "RNA",
  cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{The Seurat object that will be analyzed.}

\item{resolution_start}{The starting resolution to be used for the clustering algorithm (Louvain and Leiden algorithms).}

\item{reduction_percentage}{The amount that the starting parameter will be reduced by after each iteration (between 0 and 1).}

\item{num_clusters_start}{The starting number of clusters to be used for the clustering algorithm (K-means and Hierarchical clustering algorithms).}

\item{dims}{The dimensions to use as input features (i.e. 1:10).}

\item{algorithm}{The clustering algorithm to be used.}

\item{assay}{The assay to generate knockoffs from.}

\item{cores}{The number of cores to compute marker genes in parallel.}

\item{verbose}{Whether or not to show all logging.
Latest clustering results will be stored in the object metadata under 'callback_clusters'.
Note that 'callback_clusters' will be overwritten every time FindClustersKC is run.}
}
\value{
Returns a Seurat object where the idents have been updated with the clusters determined via the callback algorithm.
}
\description{
Given a Seurat object, returns a new Seurat that has been normalized, had variable features identified,
scaled, had principal components computed, had clusters identified, and had tSNE and UMAP embeddings determined.
}
