% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcc_fast.R
\name{FindClustersKCCFast}
\alias{FindClustersKCCFast}
\title{Runs a typical Seurat workflow on a Seurat object (up to dimensionality reduction and clustering).}
\usage{
FindClustersKCCFast(
  seurat_obj,
  resolution_start = 0.8,
  reduction_percentage = 0.2,
  num_clusters_start = 20,
  dims = 1:10,
  algorithm = "louvain",
  cores = 1
)
}
\arguments{
\item{seurat_obj}{The Seurat object that will be analyzed.}

\item{resolution_start}{The starting resolution to be used for the clustering algorithm (Louvain and Leiden algorithms).}

\item{reduction_percentage}{The amount that the starting parameter will be reduced by after each iteration (between 0 and 1).}

\item{num_clusters_start}{The starting number of clusters to be used for the clustering algorithm (K-means and Hierarchical clustering algorithms).}

\item{dims}{The dimensions to use as input features (i.e. 1:10).}

\item{algorithm}{The clustering algorithm to be used.}

\item{cores}{The number of cores to compute marker genes in parallel.}
}
\value{
Returns a Seurat object where the idents have been updated with the clusters determined via the KCC algorithm.
Latest clustering results will be stored in the object metadata under 'kcc_clusters'.
Note that 'kcc_clusters' will be overwritten every time FindClustersKC is run.
}
\description{
Given a Seurat object, returns a new Seurat that has been normalized, had variable features identified,
scaled, had principal components computed, had clusters identified, and had tSNE and UMAP embeddings determined.
}
