% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_graph.R
\name{g_difp}
\alias{g_difp}
\title{Difference of proportions plot}
\usage{
g_difp(
  data = NULL,
  method = "wilson",
  col = c("#619CFF", "#619CFF", "#F564E3"),
  ...
)
}
\arguments{
\item{data}{A \code{stat_mcr}, \code{stat_tcs} or \code{stat_ecs} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{col}{A list of colors (two colors needed).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A ggplot2 object
}
\description{
Difference of proportions plot
}
\examples{
data(origin1)
m <- callback(labour1,"offer","hist","callback","all")
s <- stat_mcr(m)
g_difp(data=s,method="student")

}
\author{
Emmanuel Duguet
}
