% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_graph.R
\name{g_ecs}
\alias{g_ecs}
\title{Exclusive callback shares plot}
\usage{
g_ecs(
  data = NULL,
  method = "wilson",
  col = c("#F564E3", "#7CAE00", "#00BFC4"),
  ...
)
}
\arguments{
\item{data}{a \code{stat_ecs} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{col}{A list of colors (three colors needed).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A ggplot2 object
}
\description{
Exclusive callback shares plot
}
\examples{
data(origin1)
m <- callback(labour1,"offer","hist","callback","all")
s <- stat_ecs(m)
g_ecs(s,method="cp")

}
\author{
Emmanuel Duguet
}
