% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_graph.R
\name{g_prop}
\alias{g_prop}
\title{Proportions' comparison plot}
\usage{
g_prop(
  data = NULL,
  method = "wilson",
  col = c("#F564E3", "#7CAE00", "#00BFC4"),
  ...
)
}
\arguments{
\item{data}{A \code{stat_mcr} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{col}{A list of colors (two colors needed).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A ggplot2 object
}
\description{
Proportions' comparison plot
}
\examples{
data(origin1)
m <- callback(labour1,"offer","hist","callback","all")
s <- stat_mcr(m)
g_prop(data=s,method="student")

}
\author{
Emmanuel Duguet
}
