% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{gender1}
\alias{gender1}
\title{Gender/Maternity discrimination (commercial and administrative jobs in the
financial sector)}
\format{
A data frame with 942 rows and 14 variables
}
\usage{
data(gender1)
}
\description{
The data were collected in January-March 2002 by Pascale Petit for her PhD
thesis.

\tabular{ll}{
\code{offer} \tab add number \cr
\code{gender}\tab Woman or Man\cr
\code{age}\tab 25 or 37 years old\cr
\code{child}\tab number of children, 0 or 3\cr
\code{educ}\tab education, BAC = Baccalauréat = A-level, BTS = 2 years of
vocational training after the A-level\cr
\code{qual}\tab qualification required by the offer, Administrative or
Commercial \cr
\code{date}\tab January 2002, February 2002 or March 2002\cr
\code{cv}\tab CV template, A or B\cr
\code{cont}\tab length of labour contract, STC = short term contract (<=1
year), LTC = long term contract (>1 year with no ending date)\cr
\code{popp}\tab promotion opportunity, Yes or No\cr
\code{train}\tab training included, Yes or No\cr
\code{negow}\tab negotiable wage, Yes or No\cr
\code{incent}\tab wage depending on output, Yes or No\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
}
}
\references{
Duguet E., Petit P., 2005. Hiring discrimination in the French financial
sector: an econometric analysis on field experiment data. Annals of
Economics and Statistics, 78: 79-102.

Petit P., 2007. The effects of age and family constraints on gender hiring
discrimination: A field experiment in the French financial sector. Labor
Economics, 14: 371-391.
}
\keyword{datasets}
