% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{gender4}
\alias{gender4}
\title{Gender/Maternity discrimination
(plumbers)}
\format{
A data frame with 1152 rows and 15 variables
}
\usage{
data(gender4)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between February and
July 2015. A candidate is defined by the variables ("gender","educ").
\itemize{
\item\bold{offer:} add number.
\item\bold{gender:} Woman or Man.
\item\bold{age:} 23 or 24 years old.
\item\bold{fname:} first name (forename).
\item\bold{lname:} last name (family name, surname).
\item\bold{educ:} education, CAP = vocational training certificate before the A
level, MAF = CAP + "One of the best French apprentice", OLY = CAP +
participation to the Worldskills Competition.
\item\bold{zip:} ZIP code.
\item\bold{cont:} length of labour contract, STC = short term contract (<=1 year),
 LTC = long term contract (>1 year with no ending date).
\item\bold{recgender:} gender of the recruiter.
\item\bold{cv:} CV template, A or B.
\item\bold{ansmode:} answering channel, email or ordinary mail ("omail").
\item\bold{sentorder:} order in which the application was sent.
\item\bold{hours:} weekly work time.
\item\bold{callback:} TRUE if there was a non negative callback.
\item\bold{date:} between February and July 2015.
}
}
\references{
Duguet E., du Parquet L., Petit P. (2022). Extracting the discrimination
components from the callback rates". TEPP Working Paper 2022-15.

Duguet, E., du Parquet, L. & Petit, P. (2022). Révéler les composantes de
 la discrimination à partir des taux de rappel. Revue française d'économie,
 XXXVII, 233-268.
}
\keyword{datasets}
