% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_graph.R
\name{graph}
\alias{graph}
\title{Generic plot function}
\usage{
graph(data, method, col, ...)
}
\arguments{
\item{data}{A \code{stat_glob}, \code{stat_mcr} or \code{stat_ecs} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{col}{A list of colors.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Generic plot function
}
