% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_graph.R
\name{graph.stat_glob}
\alias{graph.stat_glob}
\title{Global callback rates plot}
\usage{
\method{graph}{stat_glob}(
  data = NULL,
  method = "wilson",
  col = c("#F8766D", "#00BA38", "#619CFF"),
  ...
)
}
\arguments{
\item{data}{A \code{stat_glob} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{col}{A list of colors (one colour needed).}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A ggplot2 object
}
\description{
Global callback rates plot
}
\examples{
data(origin1)
m <- callback(labour1,"offer","hist","callback","all")
s <- stat_glob(m)
graph(s)

}
\author{
Emmanuel Duguet
}
