% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{inter1}
\alias{inter1}
\title{Gender/Origin discrimination (software developer)}
\format{
A data frame with 2480 rows and 11 variables
}
\usage{
data(inter1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between February and
April 2009

\tabular{ll}{
\code{offer} \tab add number \cr
\code{fname}\tab first name (forename)\cr
\code{lname}\tab last name (family name, surname)\cr
\code{gender}\tab Woman or Man\cr
\code{origin}\tab all candidates are French, the origin is suggested by the
name. F = French, S = Senegal, M = Morocco, V = Vietnam\cr
\code{date}\tab date of the application\cr
\code{sentorder}\tab order in which the application was sent\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
\code{ansorder}\tab order in which the answer was received when positive, 9
otherwise\cr
\code{cont}\tab length of labour contract, STC = short term contract (<=1
year), LTC = long term contract (>1 year with no ending date)\cr
\code{paris}\tab job located inside Paris, Yes or No\cr
}
}
\references{
Petit P., Duguet E., L'Horty Y., Du Parquet L., Sari F., 2013. Discrimination
à l'embauche : les effets du genre et de l'origine se cumulent-ils
systématiquement ? Economie et Statistique, 464-465-466: 141-153.

Duguet E., Du Parquet L, L'Horty Y., Petit P., 2015. New Evidence of Ethnic
and Gender discriminations in the French Labor Market using experimental
data: A ranking extension of responses from correspondence tests. Annals of
Economics and Statistics, 117-118: 21-39.
}
\keyword{datasets}
