% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_misc.R
\name{is.calc}
\alias{is.calc}
\title{Computational compatibility}
\usage{
is.calc(x)
}
\arguments{
\item{x}{A variable name}
}
\value{
A logical variable indicating whether the variable can be used for
computations (TRUE) or not (FALSE).
}
\description{
Computational compatibility
}
\examples{
data(origin1)
is.calc(origin1$cartime)
is.calc(origin1$callback)
is.calc(origin1$origin)

}
