% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_workflow.R
\name{jaccard}
\alias{jaccard}
\title{Returns the Jaccard coefficient of two vectors.}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
The Jaccard coefficient of two vectors.
}
\description{
Given two vectors, returns their Jaccard coefficient.
}
\examples{
jaccard(c(1,2,3), c(3,4,5))
}
