% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfwer.R
\name{knockoff.heuristic.threshold}
\alias{knockoff.heuristic.threshold}
\title{Returns the maximum of the threshold that controls the FDR at level q and the threshold that controls the k-FWER at level alpha.}
\usage{
knockoff.heuristic.threshold(W, q, k, alpha)
}
\arguments{
\item{W}{A vector of Knockoff statistics}

\item{q}{The desired control for the FDR}

\item{k}{Defines k-FWER}

\item{alpha}{The desired control for the FWER.}
}
\value{
The maximum of the threshold that controls the FDR and the threshold that controls the k-FWER
}
\description{
Given a vector of knockoff statistics, returns the maximum of the threshold that controls the FDR at
level q and the threshold that controls the k-FWER at level alpha.
}
\details{
Given \eqn{q}, \eqn{k}, and \eqn{\alpha}, calculates the threshold that controls the FDR at level \eqn{q} and the
threshold that controls the \eqn{k}-FWER at level \eqn{\alpha} and returns the maximum of the two.
}
