% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{labour1}
\alias{labour1}
\title{Labour market history discrimination
(accountants)}
\format{
A data frame with 1475 rows and 7 variables
}
\usage{
data(labour1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between February and
April 2015. A candidate is defined by the variable "hist".
\itemize{
\item\bold{offer:} add number.
\item\bold{date:} date of the application.
\item\bold{sentorder:} order in which the application was sent.
\item\bold{callback:} TRUE if there was a non negative callback.
\item\bold{cont:} length of labour contract, STC = short term contract (<=1 year),
LTC = long term contract (>1 year with no ending date).
\item\bold{paris:} job located inside Paris, Yes or No.
\item\bold{hist:} history in the labour market, LTC = Long term contract,
LTU = Long term unemployment, STU = Short term unemployment,
STC = Short term contract, PTC = Part time contract.
}
}
\references{
Duguet E., Le Gall R., L’Horty Y., Petit P., 2018. How does labour market
history influence the access to hiring interviews? International Journal
of Manpower, 39(4), 519-533.
}
\keyword{datasets}
