% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{media1}
\alias{media1}
\title{Origin/Gender discrimination and strongly negative mediatic exposure
(information technologist)}
\format{
A data frame with 3684 rows and 11 variables
}
\usage{
data(media1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between December 2008
and January 2009

\tabular{ll}{
\code{offer} \tab add number \cr
\code{date}\tab date of the application\cr
\code{sentorder}\tab order in which the application was sent\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
\code{fname}\tab first name (forename)\cr
\code{lname}\tab last name (family name, surname)\cr
\code{origin}\tab all the candidates are French, the origin is suggested by
the name, F = France, M = Morocco\cr
\code{gender}\tab Woman or Man\cr
\code{city}\tab candidate location\cr
\code{reput}\tab reputation of the city, P = privileged, U = Unprivileged\cr
\code{mediaexp}\tab strong negative mediatic exposure, Yes or No\cr
}
}
\references{
Duguet E., Gray D., L'Horty Y., Du Parquet L, Petit P., 2020. Labor market
effects of urban riots: an experimental assessment. Papers in Regional
Science, 99:787-806.
}
\keyword{datasets}
