% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{mobility1}
\alias{mobility1}
\title{Gender discrimination and mobility
(management controller)}
\format{
A data frame with 1200 rows and 12 variables
}
\usage{
data(mobility1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between October 2008
and March 2009. A candidate is defined by the variables ("gender","licenses").
\itemize{
\item \bold{offer:} add number.
\item \bold{date:} date of the application.
\item \bold{sentorder:} order in which the application was sent.
\item \bold{callback:} TRUE if there was a non negative callback.
\item \bold{ansorder:} order in which the answer was received when positive, 5
 otherwise.
\item \bold{fname:} first name (forename).
\item \bold{lname:} last name (family name, surname).
\item \bold{gender:} Woman or Man.
\item \bold{licenses:} both moto and car licenses, Yes or No.
\item \bold{cont:} length of labour contract, STC = short term contract (<=1 year),
 LTC = long term contract (>1 year with no ending date).
\item \bold{paris:} job located inside Paris, Yes or No.
\item \bold{cv:} CV template, A or B.
}
}
\references{
Duguet E., du Parquet L., L'Horty Y., Petit P., 2018. Counterproductive
hiring discrimination against women:evidence from a French correspondence
test. International Journal of Manpower, 39(1):37-50.
}
\keyword{datasets}
