% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{origin1}
\alias{origin1}
\title{Origin discrimination
(accountants)}
\format{
A data frame with 1097 rows and 20 variables
}
\usage{
data(origin1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between September and
November 2006. A candidate is defined by the variables ("nation","lnation",
"fnation"). The variables "educ" and "reput" create sample separations.
\itemize{
\item\bold{offer:} add number.
\item\bold{date:} September 2006, October 2006 or November 2006.
\item\bold{callback:} TRUE if there was a non negative callback.
\item\bold{fname:} first name (forename).
\item\bold{lname:} last name (family name, surname).
\item\bold{educ:} education, BAC = Baccalauréat = A-level, BTS = 2 years of
 vocational training after the A-level.
\item\bold{cartime:} commuting time by car (minutes).
\item\bold{cont:} length of labour contract, STC = short term contract (<=1 year),
 LTC = long term contract (>1 year with no ending date).
\item\bold{paris:} job located inside Paris, Yes or No.
\item\bold{nation:} nationality, M = Moroccan, F = French.
\item\bold{fnation:} first name sounding, M = Moroccan, F = French.
\item\bold{lnation:} last name sounding, M = Moroccan, F = French.
\item\bold{origin:} summary variable made from nation, lnation and fnation.
Example:FMF = French nationality, Moroccan family name and French first
name.
\item\bold{city:} candidate location.
\item\bold{reput:} reputation of the city, P = privileged, U = Unprivileged.
\item\bold{cv:} CV template, A or B.
\item\bold{natemp:} add obtained from the national employment agency (ANPE at the
 time of the test, France Travail today).
\item\bold{subsid:} the firm is a subsidiary of a large corporation, Yes or No.
\item\bold{ansmode:} answering channel, email or ordinary mail for all the
 applications to the same add. M = email, P = postage prepaid envelope, R =
  Marianne stamp, C = Cubitus stamp (comics character).
\item\bold{email:} answered by email by all the candidates, 1 = yes, 0 = No.
}
}
\references{
Duguet E., Leandri N., L'Horty Y., Petit P., 2010. Are young French
jobseekers of ethnic immigrant origin discriminated against? A controlled
experiment in the Paris area. Annals of Economics and Statistics, 99-100:
187-215.
}
\keyword{datasets}
