% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{origin2}
\alias{origin2}
\title{Origin discrimination (waiters)}
\format{
A data frame with 936 rows and 20 variables
}
\usage{
data(origin2)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between September and
November 2006

\tabular{ll}{
\code{offer} \tab add number \cr
\code{date}\tab September 2006, October 2006 or November 2006\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
\code{fname}\tab first name (forename)\cr
\code{lname}\tab last name (family name, surname)\cr
\code{educ}\tab education, BAC = Baccalauréat = A-level, BTS = 2 years of
vocational training after the A-level\cr
\code{cartime}\tab commuting time by car (minutes)\cr
\code{cont}\tab length of labour contract, STC = short term contract (<=1
year), LTC = long term contract (>1 year with no ending date)\cr
\code{paris}\tab job located inside Paris, Yes or No\cr
\code{nation}\tab nationality, M = Moroccan, F = French\cr
\code{fnation}\tab first name sounding, M = Moroccan, F = French\cr
\code{lnation}\tab last name sounding, M = Moroccan, F = French\cr
\code{origin}\tab summary variable made from Nation, Lnation and Fnation.
Example: FMF = French nationality, Moroccan family name and French first
name\cr
\code{city}\tab candidate location\cr
\code{reput}\tab reputation of the city, P = privileged, U = Unprivileged\cr
\code{cv}\tab CV template, A or B\cr
\code{natemp}\tab add obtained from the national employment agency (ANPE at
the time of the test, France Travail today)\cr
\code{subsid}\tab the firm is a subsidiary of a large corporation, Yes or No
\cr
\code{ansmode}\tab answering channel, email or ordinary mail for all the
applications to the same add. M = email, P = postage prepaid enveloppe, R =
Marianne stamp, C = Cubitus stamp (comics character)\cr
\code{email}\tab answered by email by all the candidates, 1 = yes, 0 = No\cr
}
}
\references{
Petit P., Duguet E., L'Horty Y., 2015. Discrimination résidentielle et
origine ethnique: une étude expérimentale sur les serveurs en Ile de France.
Economie et Prevision, 206-207: 55-69.
}
\keyword{datasets}
