% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_print.R
\name{print.callback}
\alias{print.callback}
\title{Prints the structure of the experiment}
\usage{
\method{print}{callback}(x, ...)
}
\arguments{
\item{x}{a \code{stat_count} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Computes the number of tests available for each pair of candidates
}
\examples{
data(labour1)
x <- callback(data=labour1,cluster="offer",candid="hist",callback="callback",
comp = "all")
print(x)
}
