% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_methods.R
\name{print.stat_comp}
\alias{print.stat_comp}
\title{Print the callback comparison statistics}
\usage{
\method{print}{stat_comp}(x, digits = 5, ...)
}
\arguments{
\item{x}{A \code{stat_comp} object.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Print the results of the the Student and binomial tests of equality between the callback rates.
}
\examples{
data(labour1)
print(stat_comp(callback(data=labour1, cluster="offer",candid="hist",
callback="callback", comp = "all")))
}
