% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_methods.R
\name{print.stat_count}
\alias{print.stat_count}
\title{Print the callback counts analysis}
\usage{
\method{print}{stat_count}(x, ...)
}
\arguments{
\item{x}{A \code{stat_count} object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Computes the callback numbers for each candidate.
}
\examples{
data(labour1)
print(stat_count(callback(labour1, "offer", "hist", "callback",
comp = "all")))
}
