% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_print.R
\name{print.stat_tcs}
\alias{print.stat_tcs}
\alias{print.stat_ecs}
\title{Prints the proportions of the total callback shares}
\usage{
\method{print}{stat_tcs}(x, digits = 3, ...)

\method{print}{stat_ecs}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{stat_ecs} object.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.

Printed output.
}
\description{
Prints the statistics from stat_tcs() output

Prints the statistics from stat_ecs() output
}
\examples{
data(labour1)
x <- callback(labour1, "offer", "hist", "callback", comp = "all")
print(stat_tcs(x,level=0.90))
data(labour1)
x <- callback(labour1, "offer", "hist", "callback", comp = "all")
print(stat_ecs(x,level=0.90))
}
