% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_print.R
\name{print.stat_glob}
\alias{print.stat_glob}
\title{Prints the global callback statistics}
\usage{
\method{print}{stat_glob}(x = NULL, method = "wilson", digits = 3, ...)
}
\arguments{
\item{x}{a \code{stat_glob} object.}

\item{method}{the type of confidence interval: "cp" for Clopper-Pearson,
"wilson" for Wilson (the default), "student" for Student.}

\item{digits}{minimal number of significant digits}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Prints the global callback numbers and proportions for each candidate.
}
\examples{
data(labour1)
print(stat_glob(callback(labour1, "offer", "hist", "callback",
comp = "all")))
}
