% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_print.R
\name{print.stat_mcr}
\alias{print.stat_mcr}
\title{Prints the callback proportions analysis}
\usage{
\method{print}{stat_mcr}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{stat_mcr} object.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Prints the statistics about the callback rates
}
\examples{
data(labour1)
x <- callback(labour1, "offer", "hist", "callback", comp = "all")
print(stat_mcr(x,level=0.90))
}
