% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_zipoisson.R
\name{rzipoisson}
\alias{rzipoisson}
\title{Random data generation for the zero-infalted Poisson distribution
with Poisson parameter lambda and zero proportion prop.zero.}
\usage{
rzipoisson(n, lambda, prop.zero)
}
\arguments{
\item{n}{The number of samples to be simulated.}

\item{lambda}{The Poisson rate parameter.}

\item{prop.zero}{The proportion of excess zeroes.}
}
\value{
Simulated data from ZIP(lambda, prop.zero).
}
\description{
Given the number of samples desired, a Poisson parameter,
lambda, and a zero proportion, prop.zero, simulates the number of desired
samples from ZIP(lambda, prop.zero).
}
