% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_misc.R
\name{stat_signif}
\alias{stat_signif}
\title{Significance code of a p-value}
\usage{
stat_signif(p)
}
\arguments{
\item{p}{A number between 0 and 1}
}
\value{
A character string
}
\description{
Significance code of a p-value
}
\examples{
stat_signif(0.045)

}
