% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_workflow.R
\name{compute_knockoff_filter}
\alias{compute_knockoff_filter}
\title{Returns the genes selected by the knockoff filter}
\usage{
compute_knockoff_filter(
  seurat_obj,
  cluster1,
  cluster2,
  q,
  return_all = FALSE,
  threshold = "fdr",
  num_cores = 1
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{cluster1}{The Idents of the cluster of interest in seurat_obj1}

\item{cluster2}{The Idents of the cluster of interest in seurat_obj2}

\item{q}{The desired rate to control the FDR at}

\item{return_all}{Determines if the returned object will contain all genes
or just the selected genes.}

\item{threshold}{One of "fdr", "kfwer", or "heuristic".}

\item{num_cores}{The number of cores for computing marker genes in parallel.}
}
\value{
todo
}
\description{
Given two Seurat objects, returns the  the genes selected by
the knockoff filter and their W statistics.
}
