% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_data.R
\docType{data}
\name{gender1}
\alias{gender1}
\title{Gender/Maternity discrimination
(commercial and administrative jobs in the financial sector)}
\format{
A data frame with 942 rows and 14 variables
}
\usage{
data(gender1)
}
\description{
The data were collected in January-March 2002 by Pascale Petit for her PhD
thesis (University of Paris I-Panthéon-Sorbonne, 2004). A candidate is
defined by the variables ("gender","age","child").
\itemize{
\item\bold{offer:} add number.
\item\bold{gender:} Woman or Man.
\item\bold{age:} 25 or 37 years old.
\item\bold{child:} number of children, 0 or 3.
\item\bold{educ:} education, BAC = Baccalauréat = A-level, BTS = 2 years of
vocational training after the A-level.
\item\bold{qual:} qualification required by the offer, Administrative or
Commercial.
\item\bold{date:} January 2002, February 2002 or March 2002.
\item\bold{cv:} CV template, A or B.
\item\bold{cont:} length of labour contract, STC = short term contract (<=1 year),
LTC = long term contract (>1 year with no ending date).
\item\bold{popp:} promotion opportunity, Yes or No.
\item\bold{train:} training included, Yes or No.
\item\bold{negow:} negotiable wage, Yes or No.
\item\bold{incent:} wage depending on output, Yes or No.
\item\bold{callback:} TRUE if there was a non negative callback.
}
}
\references{
Duguet E., Petit P., 2005. Hiring discrimination in the French financial
sector:an econometric analysis on field experiment data. Annals of
Economics and Statistics, 78:79-102.

Petit P., 2007. The effects of age and family constraints on gender hiring
discrimination:A field experiment in the French financial sector. Labor
Economics, 14:371-391.
}
\keyword{datasets}
