% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{labour2}
\alias{labour2}
\title{Labour market history discrimination (sales assistant)}
\format{
A data frame with 1470 rows and 7 variables
}
\usage{
data(labour2)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between January and
April 2015

\tabular{ll}{
\code{offer} \tab add number \cr
\code{date}\tab date of the application\cr
\code{sentorder}\tab order in which the application was sent\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
\code{cont}\tab length of labour contract, STC = short term contract (<=1
year), LTC = long term contract (>1 year with no ending date)\cr
\code{paris}\tab job located inside Paris, Yes or No\cr
\code{hist}\tab history in the labour market, LTC = Long term contract, LTU =
 Long term unemployment, STU = Short term unemployment, STC = Short term
 contract, PTC = Part time contract\cr
}
}
\references{
Duguet E., Le Gall R., L’Horty Y., Petit P., 2018. How does labour market
history influence the access to hiring interviews? International Journal
of Manpower, 39(4), 519-533.
}
\keyword{datasets}
