% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sets.R
\docType{data}
\name{mobility1}
\alias{mobility1}
\title{Gender discrimination and mobility (management controller)}
\format{
A data frame with 1200 rows and 12 variables
}
\usage{
data(mobility1)
}
\description{
The data were collected by the TEPP team (FR CNRS 2042) between October 2008
and March 2009

\tabular{ll}{
\code{offer} \tab add number \cr
\code{date}\tab date of the application\cr
\code{sentorder}\tab order in which the application was sent\cr
\code{callback}\tab TRUE if there was a non negative callback\cr
\code{ansorder}\tab order in which the answer was received when positive, 5
otherwise\cr
\code{fname}\tab first name (forename)\cr
\code{lname}\tab last name (family name, surname)\cr
\code{gender}\tab Woman or Man\cr
\code{licenses}\tab both moto and car licenses, Yes or No\cr
\code{cont}\tab length of labour contract, STC = short term contract (<=1
year), LTC = long term contract (>1 year with no ending date)\cr
\code{paris}\tab job located inside Paris, Yes or No\cr
\code{cv}\tab CV template, A or B\cr
}
}
\references{
Duguet E., du Parquet L., L'Horty Y., Petit P., 2018. Counterproductive
hiring discrimination against women: evidence from a French correspondence
test. International Journal of Manpower, 39(1): 37-50.
}
\keyword{datasets}
