% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback_methods.R
\name{print.stat_prop}
\alias{print.stat_prop}
\title{Print the callback proportions analysis}
\usage{
\method{print}{stat_prop}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{stat_prop} object.}

\item{digits}{minimal number of significant digits.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Printed output.
}
\description{
Computes the callbacks rates for each candidate, with their confidence intervals.
}
\examples{
data(labour1)
x <- callback(labour1, "offer", "hist", "callback", comp = "all")
print(stat_prop(x,level=0.99))
}
